reorg + icon setup
